Fixed Insert and delete
Error with image, no change with add to favorites

+++ app/src/main/java/com/jgc/primeraapplicacion/ui/moviedetail/MovieDetailPresenter.kt
@@ -1,5 +1,6 @@
 package com.jgc.primeraapplicacion.ui.moviedetail
 
+import android.util.Log
 import com.jgc.primeraapplicacion.data.local.FavoritesEntity
 import com.jgc.primeraapplicacion.data.local.LocalRepository
 import com.jgc.primeraapplicacion.data.remote.RetrofitFactory
@@ -11,7 +12,10 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
-class MovieDetailPresenter(private val view: MovieDetailView, private val localRepository: LocalRepository) {
+class MovieDetailPresenter(
+    private val view: MovieDetailView,
+    private val localRepository: LocalRepository
+) {
 
     fun init(movieId: Int) {
         val movieApi = RetrofitFactory.getMovieApi()
@@ -41,31 +45,45 @@ class MovieDetailPresenter(private val view: MovieDetailView, private val localR
 
     fun setFavorite(movieId: Int) {
         val movieApi = RetrofitFactory.getMovieApi()
+        var responseDatabase: Int? = null
         CoroutineScope(Dispatchers.IO).launch {
             val response = movieApi.getMoviesDetail(movieId, "6d247d2725f2627d9e371751ce4e8679")
             if (response.isSuccessful) {
-                  val id = response.body()!!.id
+                val id = response.body()!!.id
                 val image = response.body()!!.poster_path
                 val title = response.body()!!.title
-                val date  = response.body()!!.release_date
-                val adult  = response.body()!!.adult
-                val popularity  = response.body()!!.popularity
-                val average  = response.body()!!.vote_average
+                val date = response.body()!!.release_date
+                val adult = response.body()!!.adult
+                val popularity = response.body()!!.popularity
+                val average = response.body()!!.vote_average
                 val currentTimestamp = System.currentTimeMillis().toString()
 
+                val entity = FavoritesEntity(
+                    id,
+                    image,
+                    title,
+                    date,
+                    adult,
+                    popularity,
+                    average,
+                    currentTimestamp
+                )
 
-                val entity = FavoritesEntity(id,image,title,date,adult,popularity,average,currentTimestamp)
-
-                if (entity.id != null) {
+                responseDatabase = localRepository.checkFavorite(id)
+                Log.e("", responseDatabase.toString())
+                if (responseDatabase == null) {
                     localRepository.insertFavorite(entity)
-                }else{
-                localRepository.deleteFavorite(entity)
+                    withContext(Dispatchers.Main) {
+                        view.checkFav(responseDatabase)
+                    }
+                } else {
+                    localRepository.deleteFavorite(entity)
+                    withContext(Dispatchers.Main) {
+                        view.checkFav(responseDatabase)
+                    }
                 }
             }
 
-                withContext(Dispatchers.Main) {
-                    view.checkFav(movieId)
-                }
 
         }
     }
@@ -76,5 +94,5 @@ interface MovieDetailView {
     fun genres(genre: List<DetailGenres>)
     fun cast(cast: List<DetailCast>)
     fun crew(crew: List<DetailCast>)
-    fun checkFav(movieId: Int)
+    fun checkFav(response: Int?)
 }
